Split SessionDelegate, TaskDelegate and subclasses into global namespace.
Also removed all cases of `Foundation.` prefixes where possible by refactoring all local variables to avoid naming collisions.
