+++ Source/SessionManager.swift
@@ -24,27 +24,51 @@
 
 import Foundation
 
-/**
-    Responsible for creating and managing `Request` objects, as well as their underlying `NSURLSession`.
-*/
+/// Responsible for creating and managing `Request` objects, as well as their underlying `NSURLSession`.
 public class SessionManager {
 
-    // MARK: Properties
+    // MARK: - Helper Types
 
-    /**
-        A shared instance of `SessionManager`, used by top-level Alamofire request methods, and suitable for use 
-        directly for any ad hoc requests.
-    */
-    public static let sharedInstance: SessionManager = {
+    /// Defines whether the `MultipartFormData` encoding was successful and contains result of the encoding as
+    /// associated values.
+    ///
+    /// - Success: Represents a successful `MultipartFormData` encoding and contains the new `Request` along with
+    ///            streaming information.
+    /// - Failure: Used to represent a failure in the `MultipartFormData` encoding and also contains the encoding
+    ///            error.
+    public enum MultipartFormDataEncodingResult {
+        case success(request: Request, streamingFromDisk: Bool, streamFileURL: URL?)
+        case failure(Error)
+    }
+
+    private enum Downloadable {
+        case request(URLRequest)
+        case resumeData(Data)
+    }
+
+    private enum Uploadable {
+        case data(URLRequest, Data)
+        case file(URLRequest, URL)
+        case stream(URLRequest, InputStream)
+    }
+
+    private enum Streamable {
+        case stream(String, Int)
+        case netService(NetService)
+    }
+
+    // MARK: - Properties
+
+    /// A default instance of `SessionManager`, used by top-level Alamofire request methods, and suitable for use
+    /// directly for any ad hoc requests.
+    public static let `default`: SessionManager = {
         let configuration = URLSessionConfiguration.default
         configuration.httpAdditionalHeaders = SessionManager.defaultHTTPHeaders
 
         return SessionManager(configuration: configuration)
     }()
 
-    /**
-        Creates default values for the "Accept-Encoding", "Accept-Language" and "User-Agent" headers.
-    */
+    /// Creates default values for the "Accept-Encoding", "Accept-Language" and "User-Agent" headers.
     public static let defaultHTTPHeaders: [String: String] = {
         // Accept-Encoding HTTP Header; see https://tools.ietf.org/html/rfc7230#section-4.2.3
         let acceptEncoding: String = "gzip;q=1.0, compress;q=0.5"
@@ -98,7 +122,8 @@ public class SessionManager {
         ]
     }()
 
-    let queue = DispatchQueue(label: "Alamofire Session Manager Queue")
+    /// Default memory threshold used when encoding `MultipartFormData`.
+    public static let multipartFormDataEncodingMemoryThreshold: UInt64 = 10 * 1024 * 1024
 
     /// The underlying session.
     public let session: URLSession
@@ -109,33 +134,31 @@ public class SessionManager {
     /// Whether to start requests immediately after being constructed. `true` by default.
     public var startRequestsImmediately: Bool = true
 
-    /**
-        The background completion handler closure provided by the UIApplicationDelegate
-        `application:handleEventsForBackgroundURLSession:completionHandler:` method. By setting the background
-        completion handler, the SessionDelegate `sessionDidFinishEventsForBackgroundURLSession` closure implementation
-        will automatically call the handler.
-
-        If you need to handle your own events before the handler is called, then you need to override the
-        SessionDelegate `sessionDidFinishEventsForBackgroundURLSession` and manually call the handler when finished.
-
-        `nil` by default.
-    */
+    /// The background completion handler closure provided by the UIApplicationDelegate
+    /// `application:handleEventsForBackgroundURLSession:completionHandler:` method. By setting the background
+    /// completion handler, the SessionDelegate `sessionDidFinishEventsForBackgroundURLSession` closure implementation
+    /// will automatically call the handler.
+    ///
+    /// If you need to handle your own events before the handler is called, then you need to override the
+    /// SessionDelegate `sessionDidFinishEventsForBackgroundURLSession` and manually call the handler when finished.
+    ///
+    /// `nil` by default.
     public var backgroundCompletionHandler: (() -> Void)?
 
-    // MARK: Lifecycle
+    let queue = DispatchQueue(label: "Alamofire Session Manager Queue")
 
-    /**
-        Initializes the `Manager` instance with the specified configuration, delegate and server trust policy.
+    // MARK: - Lifecycle
 
-        - parameter configuration:            The configuration used to construct the managed session.
-                                              `NSURLSessionConfiguration.defaultSessionConfiguration()` by default.
-        - parameter delegate:                 The delegate used when initializing the session. `SessionDelegate()` by
-                                              default.
-        - parameter serverTrustPolicyManager: The server trust policy manager to use for evaluating all server trust
-                                              challenges. `nil` by default.
-
-        - returns: The new `Manager` instance.
-    */
+    /// Initializes the `SessionManager` instance with the specified configuration, delegate and server trust policy.
+    ///
+    /// - parameter configuration:            The configuration used to construct the managed session.
+    ///                                       `NSURLSessionConfiguration.defaultSessionConfiguration()` by default.
+    /// - parameter delegate:                 The delegate used when initializing the session. `SessionDelegate()` by
+    ///                                       default.
+    /// - parameter serverTrustPolicyManager: The server trust policy manager to use for evaluating all server trust
+    ///                                       challenges. `nil` by default.
+    ///
+    /// - returns: The new `SessionManager` instance.
     public init(
         configuration: URLSessionConfiguration = URLSessionConfiguration.default,
         delegate: SessionDelegate = SessionDelegate(),
@@ -147,16 +170,14 @@ public class SessionManager {
         commonInit(serverTrustPolicyManager: serverTrustPolicyManager)
     }
 
-    /**
-        Initializes the `Manager` instance with the specified session, delegate and server trust policy.
-
-        - parameter session:                  The URL session.
-        - parameter delegate:                 The delegate of the URL session. Must equal the URL session's delegate.
-        - parameter serverTrustPolicyManager: The server trust policy manager to use for evaluating all server trust
-                                              challenges. `nil` by default.
-
-        - returns: The new `Manager` instance if the URL session's delegate matches the delegate parameter.
-    */
+    /// Initializes the `SessionManager` instance with the specified session, delegate and server trust policy.
+    ///
+    /// - parameter session:                  The URL session.
+    /// - parameter delegate:                 The delegate of the URL session. Must equal the URL session's delegate.
+    /// - parameter serverTrustPolicyManager: The server trust policy manager to use for evaluating all server trust
+    ///                                       challenges. `nil` by default.
+    ///
+    /// - returns: The new `SessionManager` instance if the URL session's delegate matches the delegate parameter.
     public init?(
         session: URLSession,
         delegate: SessionDelegate,
@@ -183,43 +204,40 @@ public class SessionManager {
         session.invalidateAndCancel()
     }
 
-    // MARK: Request
+    // MARK: - Data Request
 
-    /**
-        Creates a request for the specified method, URL string, parameters, parameter encoding and headers.
-
-        - parameter method:     The HTTP method.
-        - parameter URLString:  The URL string.
-        - parameter parameters: The parameters. `nil` by default.
-        - parameter encoding:   The parameter encoding. `.URL` by default.
-        - parameter headers:    The HTTP headers. `nil` by default.
-
-        - returns: The created request.
-    */
+    /// Creates a data request for the specified method, URL string, parameters, parameter encoding and headers.
+    ///
+    /// - parameter method:     The HTTP method.
+    /// - parameter urlString:  The URL string.
+    /// - parameter parameters: The parameters. `nil` by default.
+    /// - parameter encoding:   The parameter encoding. `.URL` by default.
+    /// - parameter headers:    The HTTP headers. `nil` by default.
+    ///
+    /// - returns: The created data request.
     @discardableResult
-    public func request(
-        _ method: Method,
-        _ URLString: URLStringConvertible,
+    public func dataRequest(
+        method: Method,
+        urlString: URLStringConvertible,
         parameters: [String: AnyObject]? = nil,
         encoding: ParameterEncoding = .url,
         headers: [String: String]? = nil)
         -> Request
     {
-        let mutableURLRequest = URLRequest(method, URLString, headers: headers)
-        let encodedURLRequest = encoding.encode(mutableURLRequest, parameters: parameters).0
-        return request(encodedURLRequest)
+        let urlRequest = URLRequest(method: method, urlString: urlString, headers: headers)
+        let encodedURLRequest = encoding.encode(urlRequest, parameters: parameters).0
+
+        return dataRequest(urlRequest: encodedURLRequest)
     }
 
-    /**
-        Creates a request for the specified URL request.
-
-        If `startRequestsImmediately` is `true`, the request will have `resume()` called before being returned.
-
-        - parameter URLRequest: The URL request
-
-        - returns: The created request.
-    */
-    public func request(_ urlRequest: URLRequestConvertible) -> Request {
+    /// Creates a data request for the specified URL request.
+    ///
+    /// If `startRequestsImmediately` is `true`, the request will have `resume()` called before being returned.
+    ///
+    /// - parameter urlRequest: The URL request
+    ///
+    /// - returns: The created data request.
+    public func dataRequest(urlRequest: URLRequestConvertible) -> Request {
         var dataTask: URLSessionDataTask!
         queue.sync { dataTask = self.session.dataTask(with: urlRequest.urlRequest) }
 
@@ -232,538 +250,436 @@ public class SessionManager {
 
         return request
     }
-}
 
-// MARK: -
+    // MARK: - Download Request
 
-extension SessionManager {
-    /**
-        Responsible for handling all delegate callbacks for the underlying session.
-    */
-    public class SessionDelegate: NSObject, URLSessionDelegate, URLSessionTaskDelegate, URLSessionDataDelegate, URLSessionDownloadDelegate {
+    // MARK: URL Request
 
-        // MARK: Properties
+    /// Creates a download request for the specified method, URL string, parameters, parameter encoding, headers
+    /// and destination.
+    ///
+    /// If `startRequestsImmediately` is `true`, the request will have `resume()` called before being returned.
+    ///
+    /// - parameter method:      The HTTP method.
+    /// - parameter urlString:   The URL string.
+    /// - parameter parameters:  The parameters. `nil` by default.
+    /// - parameter encoding:    The parameter encoding. `.URL` by default.
+    /// - parameter headers:     The HTTP headers. `nil` by default.
+    /// - parameter destination: The closure used to determine the destination of the downloaded file.
+    ///
+    /// - returns: The created download request.
+    @discardableResult
+    public func downloadRequest(
+        method: Method,
+        urlString: URLStringConvertible,
+        parameters: [String: AnyObject]? = nil,
+        encoding: ParameterEncoding = .url,
+        headers: [String: String]? = nil,
+        destination: Request.DownloadFileDestination)
+        -> Request
+    {
+        let urlRequest = URLRequest(method: method, urlString: urlString, headers: headers)
+        let encodedURLRequest = encoding.encode(urlRequest, parameters: parameters).0
 
-        private var subdelegates: [Int: Request.TaskDelegate] = [:]
-        private let subdelegateQueue = DispatchQueue(label: "Alamofire Sub Delegate Queue", attributes: .concurrent)
+        return downloadRequest(urlRequest: encodedURLRequest, destination: destination)
+    }
 
-        /// Access the task delegate for the specified task in a thread-safe manner.
-        public subscript(task: URLSessionTask) -> Request.TaskDelegate? {
-            get {
-                var subdelegate: Request.TaskDelegate?
-                subdelegateQueue.sync { subdelegate = self.subdelegates[task.taskIdentifier] }
+    /// Creates a request for downloading from the specified URL request.
+    ///
+    /// If `startRequestsImmediately` is `true`, the request will have `resume()` called before being returned.
+    ///
+    /// - parameter urlRequest:  The URL request
+    /// - parameter destination: The closure used to determine the destination of the downloaded file.
+    ///
+    /// - returns: The created download request.
+    @discardableResult
+    public func downloadRequest(
+        urlRequest: URLRequestConvertible,
+        destination: Request.DownloadFileDestination)
+        -> Request
+    {
+        return downloadRequest(downloadable: .request(urlRequest.urlRequest), destination: destination)
+    }
 
-                return subdelegate
+    // MARK: Resume Data
+
+    /// Creates a request for downloading from the resume data produced from a previous request cancellation.
+    ///
+    /// If `startRequestsImmediately` is `true`, the request will have `resume()` called before being returned.
+    ///
+    /// - parameter resumeData:  The resume data. This is an opaque data blob produced by `NSURLSessionDownloadTask`
+    ///                          when a task is cancelled. See `NSURLSession -downloadTaskWithResumeData:` for
+    ///                          additional information.
+    /// - parameter destination: The closure used to determine the destination of the downloaded file.
+    ///
+    /// - returns: The created download request.
+    @discardableResult
+    public func downloadRequest(resumeData data: Data, destination: Request.DownloadFileDestination) -> Request {
+        return downloadRequest(downloadable: .resumeData(data), destination: destination)
+    }
+
+    // MARK: Private - Download Implementation
+
+    private func downloadRequest(downloadable: Downloadable, destination: Request.DownloadFileDestination) -> Request {
+        var downloadTask: URLSessionDownloadTask!
+
+        switch downloadable {
+        case .request(let request):
+            queue.sync {
+                downloadTask = self.session.downloadTask(with: request)
             }
-            set {
-                subdelegateQueue.async { self.subdelegates[task.taskIdentifier] = newValue }
+        case .resumeData(let resumeData):
+            queue.sync {
+                downloadTask = self.session.downloadTask(withResumeData: resumeData)
             }
         }
 
-        // MARK: Lifecycle
+        let request = Request(session: session, task: downloadTask)
 
-        /**
-            Initializes the `SessionDelegate` instance.
-
-            - returns: The new `SessionDelegate` instance.
-        */
-        public override init() {
-            super.init()
-        }
-
-        // MARK: NSURLSessionDelegate
-
-        /// Overrides default behavior for NSURLSessionDelegate method `URLSession:didBecomeInvalidWithError:`.
-        public var sessionDidBecomeInvalidWithError: ((Foundation.URLSession, NSError?) -> Void)?
-
-        /// Overrides default behavior for NSURLSessionDelegate method `URLSession:didReceiveChallenge:completionHandler:`.
-        public var sessionDidReceiveChallenge: ((Foundation.URLSession, URLAuthenticationChallenge) -> (Foundation.URLSession.AuthChallengeDisposition, URLCredential?))?
-
-        /// Overrides all behavior for NSURLSessionDelegate method `URLSession:didReceiveChallenge:completionHandler:` and requires the caller to call the `completionHandler`.
-        public var sessionDidReceiveChallengeWithCompletion: ((Foundation.URLSession, URLAuthenticationChallenge, (Foundation.URLSession.AuthChallengeDisposition, URLCredential?) -> Void) -> Void)?
-
-        /// Overrides default behavior for NSURLSessionDelegate method `URLSessionDidFinishEventsForBackgroundURLSession:`.
-        public var sessionDidFinishEventsForBackgroundURLSession: ((Foundation.URLSession) -> Void)?
-
-        /**
-            Tells the delegate that the session has been invalidated.
-
-            - parameter session: The session object that was invalidated.
-            - parameter error:   The error that caused invalidation, or nil if the invalidation was explicit.
-        */
-        public func urlSession(_ session: URLSession, didBecomeInvalidWithError error: Error?) {
-            sessionDidBecomeInvalidWithError?(session, error)
-        }
-
-        /**
-            Requests credentials from the delegate in response to a session-level authentication request from the remote server.
-
-            - parameter session:           The session containing the task that requested authentication.
-            - parameter challenge:         An object that contains the request for authentication.
-            - parameter completionHandler: A handler that your delegate method must call providing the disposition and credential.
-        */
-        public func urlSession(
-            _ session: URLSession,
-            didReceive challenge: URLAuthenticationChallenge,
-            completionHandler: ((Foundation.URLSession.AuthChallengeDisposition, URLCredential?) -> Void))
-        {
-            guard sessionDidReceiveChallengeWithCompletion == nil else {
-                sessionDidReceiveChallengeWithCompletion?(session, challenge, completionHandler)
-                return
+        if let downloadDelegate = request.delegate as? DownloadTaskDelegate {
+            downloadDelegate.downloadTaskDidFinishDownloadingToURL = { session, downloadTask, URL in
+                return destination(URL, downloadTask.response as! HTTPURLResponse)
             }
+        }
 
-            var disposition: Foundation.URLSession.AuthChallengeDisposition = .performDefaultHandling
-            var credential: URLCredential?
+        delegate[request.delegate.task] = request.delegate
 
-            if let sessionDidReceiveChallenge = sessionDidReceiveChallenge {
-                (disposition, credential) = sessionDidReceiveChallenge(session, challenge)
-            } else if challenge.protectionSpace.authenticationMethod == NSURLAuthenticationMethodServerTrust {
-                let host = challenge.protectionSpace.host
+        if startRequestsImmediately {
+            request.resume()
+        }
 
-                if let serverTrustPolicy = session.serverTrustPolicyManager?.serverTrustPolicyForHost(host),
-                   let serverTrust = challenge.protectionSpace.serverTrust
-                {
-                    if serverTrustPolicy.evaluateServerTrust(serverTrust, isValidForHost: host) {
-                        disposition = .useCredential
-                        credential = URLCredential(trust: serverTrust)
-                    } else {
-                        disposition = .cancelAuthenticationChallenge
+        return request
+    }
+
+    // MARK: - Upload Request
+
+    // MARK: File
+
+    /// Creates a request for uploading a file to the specified URL request.
+    ///
+    /// If `startRequestsImmediately` is `true`, the request will have `resume()` called before being returned.
+    ///
+    /// - parameter method:    The HTTP method.
+    /// - parameter urlString: The URL string.
+    /// - parameter headers:   The HTTP headers. `nil` by default.
+    /// - parameter file:      The file to upload
+    ///
+    /// - returns: The created upload request.
+    @discardableResult
+    public func uploadRequest(
+        method: Method,
+        urlString: URLStringConvertible,
+        headers: [String: String]? = nil,
+        file: URL)
+        -> Request
+    {
+        let urlRequest = URLRequest(method: method, urlString: urlString, headers: headers)
+        return uploadRequest(urlRequest: urlRequest, file: file)
+    }
+
+    /// Creates a request for uploading a file to the specified URL request.
+    ///
+    /// If `startRequestsImmediately` is `true`, the request will have `resume()` called before being returned.
+    ///
+    /// - parameter urlRequest: The URL request
+    /// - parameter file:       The file to upload
+    ///
+    /// - returns: The created upload request.
+    @discardableResult
+    public func uploadRequest(urlRequest: URLRequestConvertible, file: URL) -> Request {
+        return uploadRequest(uploadable: .file(urlRequest.urlRequest as URLRequest, file))
+    }
+
+    // MARK: Data
+
+    /// Creates a request for uploading data to the specified URL request.
+    ///
+    /// If `startRequestsImmediately` is `true`, the request will have `resume()` called before being returned.
+    ///
+    /// - parameter method:    The HTTP method.
+    /// - parameter URLString: The URL string.
+    /// - parameter headers:   The HTTP headers. `nil` by default.
+    /// - parameter data:      The data to upload
+    ///
+    /// - returns: The created upload request.
+    @discardableResult
+    public func uploadRequest(
+        method: Method,
+        urlString: URLStringConvertible,
+        headers: [String: String]? = nil,
+        data: Data)
+        -> Request
+    {
+        let urlRequest = URLRequest(method: method, urlString: urlString, headers: headers)
+        return uploadRequest(urlRequest: urlRequest, data: data)
+    }
+
+    /// Creates a request for uploading data to the specified URL request.
+    ///
+    /// If `startRequestsImmediately` is `true`, the request will have `resume()` called before being returned.
+    ///
+    /// - parameter urlRequest: The URL request.
+    /// - parameter data:       The data to upload.
+    ///
+    /// - returns: The created upload request.
+    @discardableResult
+    public func uploadRequest(urlRequest: URLRequestConvertible, data: Data) -> Request {
+        return uploadRequest(uploadable: .data(urlRequest.urlRequest as URLRequest, data))
+    }
+
+    // MARK: InputStream
+
+    /// Creates a request for uploading a stream to the specified URL request.
+    ///
+    /// If `startRequestsImmediately` is `true`, the request will have `resume()` called before being returned.
+    ///
+    /// - parameter method:    The HTTP method.
+    /// - parameter urlString: The URL string.
+    /// - parameter headers:   The HTTP headers. `nil` by default.
+    /// - parameter stream:    The stream to upload.
+    ///
+    /// - returns: The created upload request.
+    @discardableResult
+    public func uploadRequest(
+        method: Method,
+        urlString: URLStringConvertible,
+        headers: [String: String]? = nil,
+        stream: InputStream)
+        -> Request
+    {
+        let urlRequest = URLRequest(method: method, urlString: urlString, headers: headers)
+        return uploadRequest(urlRequest: urlRequest, stream: stream)
+    }
+
+    /// Creates a request for uploading a stream to the specified URL request.
+    ///
+    /// If `startRequestsImmediately` is `true`, the request will have `resume()` called before being returned.
+    ///
+    /// - parameter URLRequest: The URL request.
+    /// - parameter stream:     The stream to upload.
+    ///
+    /// - returns: The created upload request.
+    @discardableResult
+    public func uploadRequest(urlRequest: URLRequestConvertible, stream: InputStream) -> Request {
+        return uploadRequest(uploadable: .stream(urlRequest.urlRequest as URLRequest, stream))
+    }
+
+    // MARK: MultipartFormData
+
+    /// Encodes the `MultipartFormData` and creates a request to upload the result to the specified URL request.
+    ///
+    /// It is important to understand the memory implications of uploading `MultipartFormData`. If the cummulative
+    /// payload is small, encoding the data in-memory and directly uploading to a server is the by far the most
+    /// efficient approach. However, if the payload is too large, encoding the data in-memory could cause your app to
+    /// be terminated. Larger payloads must first be written to disk using input and output streams to keep the memory
+    /// footprint low, then the data can be uploaded as a stream from the resulting file. Streaming from disk MUST be
+    /// used for larger payloads such as video content.
+    ///
+    /// The `encodingMemoryThreshold` parameter allows Alamofire to automatically determine whether to encode in-memory
+    /// or stream from disk. If the content length of the `MultipartFormData` is below the `encodingMemoryThreshold`,
+    /// encoding takes place in-memory. If the content length exceeds the threshold, the data is streamed to disk
+    /// during the encoding process. Then the result is uploaded as data or as a stream depending on which encoding
+    /// technique was used.
+    ///
+    /// If `startRequestsImmediately` is `true`, the request will have `resume()` called before being returned.
+    ///
+    /// - parameter method:                  The HTTP method.
+    /// - parameter urlString:               The URL string.
+    /// - parameter headers:                 The HTTP headers. `nil` by default.
+    /// - parameter multipartFormData:       The closure used to append body parts to the `MultipartFormData`.
+    /// - parameter encodingMemoryThreshold: The encoding memory threshold in bytes.
+    ///                                      `multipartFormDataEncodingMemoryThreshold` by default.
+    /// - parameter encodingCompletion:      The closure called when the `MultipartFormData` encoding is complete.
+    public func uploadRequest(
+        method: Method,
+        urlString: URLStringConvertible,
+        headers: [String: String]? = nil,
+        multipartFormData: (MultipartFormData) -> Void,
+        encodingMemoryThreshold: UInt64 = SessionManager.multipartFormDataEncodingMemoryThreshold,
+        encodingCompletion: ((MultipartFormDataEncodingResult) -> Void)?)
+    {
+        let urlRequest = URLRequest(method: method, urlString: urlString, headers: headers)
+
+        return uploadRequest(
+            urlRequest: urlRequest,
+            multipartFormData: multipartFormData,
+            encodingMemoryThreshold: encodingMemoryThreshold,
+            encodingCompletion: encodingCompletion
+        )
+    }
+
+    /// Encodes the `MultipartFormData` and creates a request to upload the result to the specified URL request.
+    ///
+    /// It is important to understand the memory implications of uploading `MultipartFormData`. If the cummulative
+    /// payload is small, encoding the data in-memory and directly uploading to a server is the by far the most
+    /// efficient approach. However, if the payload is too large, encoding the data in-memory could cause your app to
+    /// be terminated. Larger payloads must first be written to disk using input and output streams to keep the memory
+    /// footprint low, then the data can be uploaded as a stream from the resulting file. Streaming from disk MUST be
+    /// used for larger payloads such as video content.
+    ///
+    /// The `encodingMemoryThreshold` parameter allows Alamofire to automatically determine whether to encode in-memory
+    /// or stream from disk. If the content length of the `MultipartFormData` is below the `encodingMemoryThreshold`,
+    /// encoding takes place in-memory. If the content length exceeds the threshold, the data is streamed to disk
+    /// during the encoding process. Then the result is uploaded as data or as a stream depending on which encoding
+    /// technique was used.
+    ///
+    /// If `startRequestsImmediately` is `true`, the request will have `resume()` called before being returned.
+    ///
+    /// - parameter urlRequest:              The URL request.
+    /// - parameter multipartFormData:       The closure used to append body parts to the `MultipartFormData`.
+    /// - parameter encodingMemoryThreshold: The encoding memory threshold in bytes.
+    ///                                      `multipartFormDataEncodingMemoryThreshold` by default.
+    /// - parameter encodingCompletion:      The closure called when the `MultipartFormData` encoding is complete.
+    public func uploadRequest(
+        urlRequest: URLRequestConvertible,
+        multipartFormData: (MultipartFormData) -> Void,
+        encodingMemoryThreshold: UInt64 = SessionManager.multipartFormDataEncodingMemoryThreshold,
+        encodingCompletion: ((MultipartFormDataEncodingResult) -> Void)?)
+    {
+        DispatchQueue.global(qos: .utility).async {
+            let formData = MultipartFormData()
+            multipartFormData(formData)
+
+            var urlRequestWithContentType = urlRequest.urlRequest
+            urlRequestWithContentType.setValue(formData.contentType, forHTTPHeaderField: "Content-Type")
+
+            let isBackgroundSession = self.session.configuration.identifier != nil
+
+            if formData.contentLength < encodingMemoryThreshold && !isBackgroundSession {
+                do {
+                    let data = try formData.encode()
+
+                    let encodingResult = MultipartFormDataEncodingResult.success(
+                        request: self.uploadRequest(urlRequest: urlRequestWithContentType, data: data),
+                        streamingFromDisk: false,
+                        streamFileURL: nil
+                    )
+
+                    DispatchQueue.main.async { encodingCompletion?(encodingResult) }
+                } catch {
+                    DispatchQueue.main.async { encodingCompletion?(.failure(error as NSError)) }
+                }
+            } else {
+                let fileManager = FileManager.default
+                let tempDirectoryURL = URL(fileURLWithPath: NSTemporaryDirectory())
+                let directoryURL = tempDirectoryURL.appendingPathComponent("org.alamofire.manager/multipart.form.data")
+                let fileName = UUID().uuidString
+                let fileURL = directoryURL.appendingPathComponent(fileName)
+
+                do {
+                    try fileManager.createDirectory(at: directoryURL, withIntermediateDirectories: true, attributes: nil)
+                    try formData.writeEncodedDataToDisk(fileURL)
+
+                    DispatchQueue.main.async {
+                        let encodingResult = MultipartFormDataEncodingResult.success(
+                            request: self.uploadRequest(urlRequest: urlRequestWithContentType, file: fileURL),
+                            streamingFromDisk: true,
+                            streamFileURL: fileURL
+                        )
+                        encodingCompletion?(encodingResult)
                     }
+                } catch {
+                    DispatchQueue.main.async { encodingCompletion?(.failure(error as NSError)) }
                 }
-            }
-
-            completionHandler(disposition, credential)
-        }
-
-        /**
-            Tells the delegate that all messages enqueued for a session have been delivered.
-
-            - parameter session: The session that no longer has any outstanding requests.
-        */
-#if !os(OSX)
-        public func urlSessionDidFinishEvents(forBackgroundURLSession session: URLSession) {
-            sessionDidFinishEventsForBackgroundURLSession?(session)
-        }
-#endif
-
-        // MARK: NSURLSessionTaskDelegate
-
-        /// Overrides default behavior for NSURLSessionTaskDelegate method `URLSession:task:willPerformHTTPRedirection:newRequest:completionHandler:`.
-        public var taskWillPerformHTTPRedirection: ((Foundation.URLSession, URLSessionTask, HTTPURLResponse, Foundation.URLRequest) -> Foundation.URLRequest?)?
-
-        /// Overrides all behavior for NSURLSessionTaskDelegate method `URLSession:task:willPerformHTTPRedirection:newRequest:completionHandler:` and
-        /// requires the caller to call the `completionHandler`.
-        public var taskWillPerformHTTPRedirectionWithCompletion: ((Foundation.URLSession, URLSessionTask, HTTPURLResponse, Foundation.URLRequest, (Foundation.URLRequest?) -> Void) -> Void)?
-
-        /// Overrides default behavior for NSURLSessionTaskDelegate method `URLSession:task:didReceiveChallenge:completionHandler:`.
-        public var taskDidReceiveChallenge: ((Foundation.URLSession, URLSessionTask, URLAuthenticationChallenge) -> (Foundation.URLSession.AuthChallengeDisposition, URLCredential?))?
-
-        /// Overrides all behavior for NSURLSessionTaskDelegate method `URLSession:task:didReceiveChallenge:completionHandler:` and
-        /// requires the caller to call the `completionHandler`.
-        public var taskDidReceiveChallengeWithCompletion: ((Foundation.URLSession, URLSessionTask, URLAuthenticationChallenge, (Foundation.URLSession.AuthChallengeDisposition, URLCredential?) -> Void) -> Void)?
-
-        /// Overrides default behavior for NSURLSessionTaskDelegate method `URLSession:session:task:needNewBodyStream:`.
-        public var taskNeedNewBodyStream: ((Foundation.URLSession, URLSessionTask) -> InputStream?)?
-
-        /// Overrides all behavior for NSURLSessionTaskDelegate method `URLSession:session:task:needNewBodyStream:` and
-        /// requires the caller to call the `completionHandler`.
-        public var taskNeedNewBodyStreamWithCompletion: ((Foundation.URLSession, URLSessionTask, (InputStream?) -> Void) -> Void)?
-
-        /// Overrides default behavior for NSURLSessionTaskDelegate method `URLSession:task:didSendBodyData:totalBytesSent:totalBytesExpectedToSend:`.
-        public var taskDidSendBodyData: ((Foundation.URLSession, URLSessionTask, Int64, Int64, Int64) -> Void)?
-
-        /// Overrides default behavior for NSURLSessionTaskDelegate method `URLSession:task:didCompleteWithError:`.
-        public var taskDidComplete: ((Foundation.URLSession, URLSessionTask, NSError?) -> Void)?
-
-        /**
-            Tells the delegate that the remote server requested an HTTP redirect.
-
-            - parameter session:           The session containing the task whose request resulted in a redirect.
-            - parameter task:              The task whose request resulted in a redirect.
-            - parameter response:          An object containing the server’s response to the original request.
-            - parameter request:           A URL request object filled out with the new location.
-            - parameter completionHandler: A closure that your handler should call with either the value of the request
-                                           parameter, a modified URL request object, or NULL to refuse the redirect and
-                                           return the body of the redirect response.
-        */
-        public func urlSession(
-            _ session: URLSession,
-            task: URLSessionTask,
-            willPerformHTTPRedirection response: HTTPURLResponse,
-            newRequest request: URLRequest,
-            completionHandler: (URLRequest?) -> Void)
-        {
-            guard taskWillPerformHTTPRedirectionWithCompletion == nil else {
-                taskWillPerformHTTPRedirectionWithCompletion?(session, task, response, request, completionHandler)
-                return
-            }
-
-            var redirectRequest: Foundation.URLRequest? = request
-
-            if let taskWillPerformHTTPRedirection = taskWillPerformHTTPRedirection {
-                redirectRequest = taskWillPerformHTTPRedirection(session, task, response, request)
-            }
-
-            completionHandler(redirectRequest)
-        }
-
-        /**
-            Requests credentials from the delegate in response to an authentication request from the remote server.
-
-            - parameter session:           The session containing the task whose request requires authentication.
-            - parameter task:              The task whose request requires authentication.
-            - parameter challenge:         An object that contains the request for authentication.
-            - parameter completionHandler: A handler that your delegate method must call providing the disposition and credential.
-        */
-        public func urlSession(
-            _ session: URLSession,
-            task: URLSessionTask,
-            didReceive challenge: URLAuthenticationChallenge,
-            completionHandler: (Foundation.URLSession.AuthChallengeDisposition, URLCredential?) -> Void)
-        {
-            guard taskDidReceiveChallengeWithCompletion == nil else {
-                taskDidReceiveChallengeWithCompletion?(session, task, challenge, completionHandler)
-                return
-            }
-
-            if let taskDidReceiveChallenge = taskDidReceiveChallenge {
-                let result = taskDidReceiveChallenge(session, task, challenge)
-                completionHandler(result.0, result.1)
-            } else if let delegate = self[task] {
-                delegate.urlSession(
-                    session,
-                    task: task,
-                    didReceive: challenge,
-                    completionHandler: completionHandler
-                )
-            } else {
-                urlSession(session, didReceive: challenge, completionHandler: completionHandler)
-            }
-        }
-
-        /**
-            Tells the delegate when a task requires a new request body stream to send to the remote server.
-
-            - parameter session:           The session containing the task that needs a new body stream.
-            - parameter task:              The task that needs a new body stream.
-            - parameter completionHandler: A completion handler that your delegate method should call with the new body stream.
-        */
-        public func urlSession(
-            _ session: URLSession,
-            task: URLSessionTask,
-            needNewBodyStream completionHandler: (InputStream?) -> Void)
-        {
-            guard taskNeedNewBodyStreamWithCompletion == nil else {
-                taskNeedNewBodyStreamWithCompletion?(session, task, completionHandler)
-                return
-            }
-
-            if let taskNeedNewBodyStream = taskNeedNewBodyStream {
-                completionHandler(taskNeedNewBodyStream(session, task))
-            } else if let delegate = self[task] {
-                delegate.urlSession(session, task: task, needNewBodyStream: completionHandler)
-            }
-        }
-
-        /**
-            Periodically informs the delegate of the progress of sending body content to the server.
-
-            - parameter session:                  The session containing the data task.
-            - parameter task:                     The data task.
-            - parameter bytesSent:                The number of bytes sent since the last time this delegate method was called.
-            - parameter totalBytesSent:           The total number of bytes sent so far.
-            - parameter totalBytesExpectedToSend: The expected length of the body data.
-        */
-        public func urlSession(
-            _ session: URLSession,
-            task: URLSessionTask,
-            didSendBodyData bytesSent: Int64,
-            totalBytesSent: Int64,
-            totalBytesExpectedToSend: Int64)
-        {
-            if let taskDidSendBodyData = taskDidSendBodyData {
-                taskDidSendBodyData(session, task, bytesSent, totalBytesSent, totalBytesExpectedToSend)
-            } else if let delegate = self[task] as? Request.UploadTaskDelegate {
-                delegate.URLSession(
-                    session,
-                    task: task,
-                    didSendBodyData: bytesSent,
-                    totalBytesSent: totalBytesSent,
-                    totalBytesExpectedToSend: totalBytesExpectedToSend
-                )
-            }
-        }
-
-        /**
-            Tells the delegate that the task finished transferring data.
-
-            - parameter session: The session containing the task whose request finished transferring data.
-            - parameter task:    The task whose request finished transferring data.
-            - parameter error:   If an error occurred, an error object indicating how the transfer failed, otherwise nil.
-        */
-        public func urlSession(_ session: URLSession, task: URLSessionTask, didCompleteWithError error: Error?) {
-            if let taskDidComplete = taskDidComplete {
-                taskDidComplete(session, task, error)
-            } else if let delegate = self[task] {
-                delegate.urlSession(session, task: task, didCompleteWithError: error)
-            }
-
-            NotificationCenter.default.post(
-                name: Notification.Name.Task.DidComplete,
-                object: self,
-                userInfo: [Notification.Key.Task: task]
-            )
-
-            self[task] = nil
-        }
-
-        // MARK: NSURLSessionDataDelegate
-
-        /// Overrides default behavior for NSURLSessionDataDelegate method `URLSession:dataTask:didReceiveResponse:completionHandler:`.
-        public var dataTaskDidReceiveResponse: ((Foundation.URLSession, URLSessionDataTask, URLResponse) -> Foundation.URLSession.ResponseDisposition)?
-
-        /// Overrides all behavior for NSURLSessionDataDelegate method `URLSession:dataTask:didReceiveResponse:completionHandler:` and
-        /// requires caller to call the `completionHandler`.
-        public var dataTaskDidReceiveResponseWithCompletion: ((Foundation.URLSession, URLSessionDataTask, URLResponse, (Foundation.URLSession.ResponseDisposition) -> Void) -> Void)?
-
-        /// Overrides default behavior for NSURLSessionDataDelegate method `URLSession:dataTask:didBecomeDownloadTask:`.
-        public var dataTaskDidBecomeDownloadTask: ((Foundation.URLSession, URLSessionDataTask, URLSessionDownloadTask) -> Void)?
-
-        /// Overrides default behavior for NSURLSessionDataDelegate method `URLSession:dataTask:didReceiveData:`.
-        public var dataTaskDidReceiveData: ((Foundation.URLSession, URLSessionDataTask, Data) -> Void)?
-
-        /// Overrides default behavior for NSURLSessionDataDelegate method `URLSession:dataTask:willCacheResponse:completionHandler:`.
-        public var dataTaskWillCacheResponse: ((Foundation.URLSession, URLSessionDataTask, CachedURLResponse) -> CachedURLResponse?)?
-
-        /// Overrides all behavior for NSURLSessionDataDelegate method `URLSession:dataTask:willCacheResponse:completionHandler:` and
-        /// requires caller to call the `completionHandler`.
-        public var dataTaskWillCacheResponseWithCompletion: ((Foundation.URLSession, URLSessionDataTask, CachedURLResponse, (CachedURLResponse?) -> Void) -> Void)?
-
-        /**
-            Tells the delegate that the data task received the initial reply (headers) from the server.
-
-            - parameter session:           The session containing the data task that received an initial reply.
-            - parameter dataTask:          The data task that received an initial reply.
-            - parameter response:          A URL response object populated with headers.
-            - parameter completionHandler: A completion handler that your code calls to continue the transfer, passing a
-                                           constant to indicate whether the transfer should continue as a data task or
-                                           should become a download task.
-        */
-        public func urlSession(
-            _ session: URLSession,
-            dataTask: URLSessionDataTask,
-            didReceive response: URLResponse,
-            completionHandler: (URLSession.ResponseDisposition) -> Void)
-        {
-            guard dataTaskDidReceiveResponseWithCompletion == nil else {
-                dataTaskDidReceiveResponseWithCompletion?(session, dataTask, response, completionHandler)
-                return
-            }
-
-            var disposition: Foundation.URLSession.ResponseDisposition = .allow
-
-            if let dataTaskDidReceiveResponse = dataTaskDidReceiveResponse {
-                disposition = dataTaskDidReceiveResponse(session, dataTask, response)
-            }
-
-            completionHandler(disposition)
-        }
-
-        /**
-            Tells the delegate that the data task was changed to a download task.
-
-            - parameter session:      The session containing the task that was replaced by a download task.
-            - parameter dataTask:     The data task that was replaced by a download task.
-            - parameter downloadTask: The new download task that replaced the data task.
-        */
-        public func urlSession(
-            _ session: URLSession,
-            dataTask: URLSessionDataTask,
-            didBecome downloadTask: URLSessionDownloadTask)
-        {
-            if let dataTaskDidBecomeDownloadTask = dataTaskDidBecomeDownloadTask {
-                dataTaskDidBecomeDownloadTask(session, dataTask, downloadTask)
-            } else {
-                let downloadDelegate = Request.DownloadTaskDelegate(task: downloadTask)
-                self[downloadTask] = downloadDelegate
-            }
-        }
-
-        /**
-            Tells the delegate that the data task has received some of the expected data.
-
-            - parameter session:  The session containing the data task that provided data.
-            - parameter dataTask: The data task that provided data.
-            - parameter data:     A data object containing the transferred data.
-        */
-        public func urlSession(_ session: URLSession, dataTask: URLSessionDataTask, didReceive data: Data) {
-            if let dataTaskDidReceiveData = dataTaskDidReceiveData {
-                dataTaskDidReceiveData(session, dataTask, data)
-            } else if let delegate = self[dataTask] as? Request.DataTaskDelegate {
-                delegate.urlSession(session, dataTask: dataTask, didReceive: data)
-            }
-        }
-
-        /**
-            Asks the delegate whether the data (or upload) task should store the response in the cache.
-
-            - parameter session:           The session containing the data (or upload) task.
-            - parameter dataTask:          The data (or upload) task.
-            - parameter proposedResponse:  The default caching behavior. This behavior is determined based on the current
-                                           caching policy and the values of certain received headers, such as the Pragma
-                                           and Cache-Control headers.
-            - parameter completionHandler: A block that your handler must call, providing either the original proposed
-                                           response, a modified version of that response, or NULL to prevent caching the
-                                           response. If your delegate implements this method, it must call this completion
-                                           handler; otherwise, your app leaks memory.
-        */
-        public func urlSession(
-            _ session: URLSession,
-            dataTask: URLSessionDataTask,
-            willCacheResponse proposedResponse: CachedURLResponse,
-            completionHandler: (CachedURLResponse?) -> Void)
-        {
-            guard dataTaskWillCacheResponseWithCompletion == nil else {
-                dataTaskWillCacheResponseWithCompletion?(session, dataTask, proposedResponse, completionHandler)
-                return
-            }
-
-            if let dataTaskWillCacheResponse = dataTaskWillCacheResponse {
-                completionHandler(dataTaskWillCacheResponse(session, dataTask, proposedResponse))
-            } else if let delegate = self[dataTask] as? Request.DataTaskDelegate {
-                delegate.urlSession(
-                    session,
-                    dataTask: dataTask,
-                    willCacheResponse: proposedResponse,
-                    completionHandler: completionHandler
-                )
-            } else {
-                completionHandler(proposedResponse)
-            }
-        }
-
-        // MARK: NSURLSessionDownloadDelegate
-
-        /// Overrides default behavior for NSURLSessionDownloadDelegate method `URLSession:downloadTask:didFinishDownloadingToURL:`.
-        public var downloadTaskDidFinishDownloadingToURL: ((Foundation.URLSession, URLSessionDownloadTask, URL) -> Void)?
-
-        /// Overrides default behavior for NSURLSessionDownloadDelegate method `URLSession:downloadTask:didWriteData:totalBytesWritten:totalBytesExpectedToWrite:`.
-        public var downloadTaskDidWriteData: ((Foundation.URLSession, URLSessionDownloadTask, Int64, Int64, Int64) -> Void)?
-
-        /// Overrides default behavior for NSURLSessionDownloadDelegate method `URLSession:downloadTask:didResumeAtOffset:expectedTotalBytes:`.
-        public var downloadTaskDidResumeAtOffset: ((Foundation.URLSession, URLSessionDownloadTask, Int64, Int64) -> Void)?
-
-        /**
-            Tells the delegate that a download task has finished downloading.
-
-            - parameter session:      The session containing the download task that finished.
-            - parameter downloadTask: The download task that finished.
-            - parameter location:     A file URL for the temporary file. Because the file is temporary, you must either
-                                      open the file for reading or move it to a permanent location in your app’s sandbox
-                                      container directory before returning from this delegate method.
-        */
-        public func urlSession(
-            _ session: URLSession,
-            downloadTask: URLSessionDownloadTask,
-            didFinishDownloadingTo location: URL)
-        {
-            if let downloadTaskDidFinishDownloadingToURL = downloadTaskDidFinishDownloadingToURL {
-                downloadTaskDidFinishDownloadingToURL(session, downloadTask, location)
-            } else if let delegate = self[downloadTask] as? Request.DownloadTaskDelegate {
-                delegate.urlSession(session, downloadTask: downloadTask, didFinishDownloadingTo: location)
-            }
-        }
-
-        /**
-            Periodically informs the delegate about the download’s progress.
-
-            - parameter session:                   The session containing the download task.
-            - parameter downloadTask:              The download task.
-            - parameter bytesWritten:              The number of bytes transferred since the last time this delegate
-                                                   method was called.
-            - parameter totalBytesWritten:         The total number of bytes transferred so far.
-            - parameter totalBytesExpectedToWrite: The expected length of the file, as provided by the Content-Length
-                                                   header. If this header was not provided, the value is
-                                                   `NSURLSessionTransferSizeUnknown`.
-        */
-        public func urlSession(
-            _ session: URLSession,
-            downloadTask: URLSessionDownloadTask,
-            didWriteData bytesWritten: Int64,
-            totalBytesWritten: Int64,
-            totalBytesExpectedToWrite: Int64)
-        {
-            if let downloadTaskDidWriteData = downloadTaskDidWriteData {
-                downloadTaskDidWriteData(session, downloadTask, bytesWritten, totalBytesWritten, totalBytesExpectedToWrite)
-            } else if let delegate = self[downloadTask] as? Request.DownloadTaskDelegate {
-                delegate.urlSession(
-                    session,
-                    downloadTask: downloadTask,
-                    didWriteData: bytesWritten,
-                    totalBytesWritten: totalBytesWritten,
-                    totalBytesExpectedToWrite: totalBytesExpectedToWrite
-                )
-            }
-        }
-
-        /**
-            Tells the delegate that the download task has resumed downloading.
-
-            - parameter session:            The session containing the download task that finished.
-            - parameter downloadTask:       The download task that resumed. See explanation in the discussion.
-            - parameter fileOffset:         If the file's cache policy or last modified date prevents reuse of the
-                                            existing content, then this value is zero. Otherwise, this value is an
-                                            integer representing the number of bytes on disk that do not need to be
-                                            retrieved again.
-            - parameter expectedTotalBytes: The expected length of the file, as provided by the Content-Length header.
-                                            If this header was not provided, the value is NSURLSessionTransferSizeUnknown.
-        */
-        public func urlSession(
-            _ session: URLSession,
-            downloadTask: URLSessionDownloadTask,
-            didResumeAtOffset fileOffset: Int64,
-            expectedTotalBytes: Int64)
-        {
-            if let downloadTaskDidResumeAtOffset = downloadTaskDidResumeAtOffset {
-                downloadTaskDidResumeAtOffset(session, downloadTask, fileOffset, expectedTotalBytes)
-            } else if let delegate = self[downloadTask] as? Request.DownloadTaskDelegate {
-                delegate.urlSession(
-                    session,
-                    downloadTask: downloadTask,
-                    didResumeAtOffset: fileOffset,
-                    expectedTotalBytes: expectedTotalBytes
-                )
-            }
-        }
-
-        // MARK: NSURLSessionStreamDelegate
-
-        var _streamTaskReadClosed: Any?
-        var _streamTaskWriteClosed: Any?
-        var _streamTaskBetterRouteDiscovered: Any?
-        var _streamTaskDidBecomeInputStream: Any?
-
-        // MARK: NSObject
-
-        public override func responds(to selector: Selector) -> Bool {
-            #if !os(OSX)
-                if selector == #selector(URLSessionDelegate.urlSessionDidFinishEvents(forBackgroundURLSession:)) {
-                    return sessionDidFinishEventsForBackgroundURLSession != nil
-                }
-            #endif
-
-            switch selector {
-            case #selector(URLSessionDelegate.urlSession(_:didBecomeInvalidWithError:)):
-                return sessionDidBecomeInvalidWithError != nil
-            case #selector(URLSessionDelegate.urlSession(_:didReceive:completionHandler:)):
-                return (sessionDidReceiveChallenge != nil  || sessionDidReceiveChallengeWithCompletion != nil)
-            case #selector(URLSessionTaskDelegate.urlSession(_:task:willPerformHTTPRedirection:newRequest:completionHandler:)):
-                return (taskWillPerformHTTPRedirection != nil || taskWillPerformHTTPRedirectionWithCompletion != nil)
-            case #selector(URLSessionDataDelegate.urlSession(_:dataTask:didReceive:completionHandler:)):
-                return (dataTaskDidReceiveResponse != nil || dataTaskDidReceiveResponseWithCompletion != nil)
-            default:
-                return self.dynamicType.instancesRespond(to: selector)
             }
         }
     }
+
+    // MARK: Private - Upload Implementation
+
+    private func uploadRequest(uploadable: Uploadable) -> Request {
+        var uploadTask: URLSessionUploadTask!
+        var HTTPBodyStream: InputStream?
+
+        switch uploadable {
+        case .data(let request, let data):
+            queue.sync {
+                uploadTask = self.session.uploadTask(with: request, from: data)
+            }
+        case .file(let request, let fileURL):
+            queue.sync {
+                uploadTask = self.session.uploadTask(with: request, fromFile: fileURL)
+            }
+        case .stream(let request, let stream):
+            queue.sync {
+                uploadTask = self.session.uploadTask(withStreamedRequest: request)
+            }
+
+            HTTPBodyStream = stream
+        }
+
+        let request = Request(session: session, task: uploadTask)
+
+        if HTTPBodyStream != nil {
+            request.delegate.taskNeedNewBodyStream = { _, _ in
+                return HTTPBodyStream
+            }
+        }
+
+        delegate[request.delegate.task] = request.delegate
+
+        if startRequestsImmediately {
+            request.resume()
+        }
+
+        return request
+    }
+
+#if !os(watchOS)
+
+    // MARK: - Stream Request
+
+    // MARK: Hostname and Port
+
+    /// Creates a stream request for bidirectional streaming with the given hostname and port.
+    ///
+    /// - parameter hostName: The hostname of the server to connect to.
+    /// - parameter port:     The port of the server to connect to.
+    ///
+    /// - returns: The created stream request.
+    @discardableResult
+    public func streamRequest(hostName: String, port: Int) -> Request {
+        return streamRequest(streamable: .stream(hostName, port))
+    }
+
+    // MARK: NetService
+
+    /// Creates a request for bidirectional streaming with the given `NSNetService`.
+    ///
+    /// - parameter netService: The net service used to identify the endpoint.
+    ///
+    /// - returns: The created stream request.
+    @discardableResult
+    public func streamRequest(netService: NetService) -> Request {
+        return streamRequest(streamable: .netService(netService))
+    }
+
+    // MARK: Private - Stream Implementation
+
+    private func streamRequest(streamable: Streamable) -> Request {
+        var streamTask: URLSessionStreamTask!
+
+        switch streamable {
+        case .stream(let hostName, let port):
+            queue.sync {
+                streamTask = self.session.streamTask(withHostName: hostName, port: port)
+            }
+        case .netService(let netService):
+            queue.sync {
+                streamTask = self.session.streamTask(with: netService)
+            }
+        }
+
+        let request = Request(session: session, task: streamTask)
+
+        delegate[request.delegate.task] = request.delegate
+
+        if startRequestsImmediately {
+            request.resume()
+        }
+
+        return request
+    }
+
+#endif
 }
