+++ Tests/ParameterEncodingTests.swift
@@ -27,7 +27,7 @@ import Foundation
 import XCTest
 
 class ParameterEncodingTestCase: BaseTestCase {
-    let urlRequest = Foundation.URLRequest(url: URL(string: "https://example.com/")!)
+    let urlRequest = URLRequest(url: URL(string: "https://example.com/")!)
 }
 
 // MARK: -
@@ -40,10 +40,10 @@ class URLParameterEncodingTestCase: ParameterEncodingTestCase {
     func testURLParameterEncodeNilParameters() {
         // Given
         // When
-        let (URLRequest, _) = encoding.encode(self.urlRequest, parameters: nil)
+        let (urlRequest, _) = encoding.encode(self.urlRequest, parameters: nil)
 
         // Then
-        XCTAssertNil(URLRequest.url?.query, "query should be nil")
+        XCTAssertNil(urlRequest.url?.query, "query should be nil")
     }
 
     func testURLParameterEncodeEmptyDictionaryParameter() {
@@ -51,10 +51,10 @@ class URLParameterEncodingTestCase: ParameterEncodingTestCase {
         let parameters: [String: AnyObject] = [:]
 
         // When
-        let (URLRequest, _) = encoding.encode(self.urlRequest, parameters: parameters)
+        let (urlRequest, _) = encoding.encode(self.urlRequest, parameters: parameters)
 
         // Then
-        XCTAssertNil(URLRequest.url?.query, "query should be nil")
+        XCTAssertNil(urlRequest.url?.query, "query should be nil")
     }
 
     func testURLParameterEncodeOneStringKeyStringValueParameter() {
@@ -62,26 +62,26 @@ class URLParameterEncodingTestCase: ParameterEncodingTestCase {
         let parameters = ["foo": "bar"]
 
         // When
-        let (URLRequest, _) = encoding.encode(self.urlRequest, parameters: parameters)
+        let (urlRequest, _) = encoding.encode(self.urlRequest, parameters: parameters)
 
         // Then
-        XCTAssertEqual(URLRequest.url?.query ?? "", "foo=bar", "query is incorrect")
+        XCTAssertEqual(urlRequest.url?.query ?? "", "foo=bar", "query is incorrect")
     }
 
     func testURLParameterEncodeOneStringKeyStringValueParameterAppendedToQuery() {
         // Given
         var mutableURLRequest = self.urlRequest.urlRequest
-        var urlComponents = Foundation.URLComponents(url: mutableURLRequest.url!, resolvingAgainstBaseURL: false)!
+        var urlComponents = URLComponents(url: mutableURLRequest.url!, resolvingAgainstBaseURL: false)!
         urlComponents.query = "baz=qux"
         mutableURLRequest.url = urlComponents.url
 
         let parameters = ["foo": "bar"]
 
         // When
-        let (URLRequest, _) = encoding.encode(mutableURLRequest, parameters: parameters)
+        let (urlRequest, _) = encoding.encode(mutableURLRequest, parameters: parameters)
 
         // Then
-        XCTAssertEqual(URLRequest.url?.query ?? "", "baz=qux&foo=bar", "query is incorrect")
+        XCTAssertEqual(urlRequest.url?.query ?? "", "baz=qux&foo=bar", "query is incorrect")
     }
 
     func testURLParameterEncodeTwoStringKeyStringValueParameters() {
@@ -89,10 +89,10 @@ class URLParameterEncodingTestCase: ParameterEncodingTestCase {
         let parameters = ["foo": "bar", "baz": "qux"]
 
         // When
-        let (URLRequest, _) = encoding.encode(self.urlRequest, parameters: parameters)
+        let (urlRequest, _) = encoding.encode(self.urlRequest, parameters: parameters)
 
         // Then
-        XCTAssertEqual(URLRequest.url?.query ?? "", "baz=qux&foo=bar", "query is incorrect")
+        XCTAssertEqual(urlRequest.url?.query ?? "", "baz=qux&foo=bar", "query is incorrect")
     }
 
     func testURLParameterEncodeStringKeyIntegerValueParameter() {
@@ -100,10 +100,10 @@ class URLParameterEncodingTestCase: ParameterEncodingTestCase {
         let parameters = ["foo": 1]
 
         // When
-        let (URLRequest, _) = encoding.encode(self.urlRequest, parameters: parameters)
+        let (urlRequest, _) = encoding.encode(self.urlRequest, parameters: parameters)
 
         // Then
-        XCTAssertEqual(URLRequest.url?.query ?? "", "foo=1", "query is incorrect")
+        XCTAssertEqual(urlRequest.url?.query ?? "", "foo=1", "query is incorrect")
     }
 
     func testURLParameterEncodeStringKeyDoubleValueParameter() {
@@ -111,10 +111,10 @@ class URLParameterEncodingTestCase: ParameterEncodingTestCase {
         let parameters = ["foo": 1.1]
 
         // When
-        let (URLRequest, _) = encoding.encode(self.urlRequest, parameters: parameters)
+        let (urlRequest, _) = encoding.encode(self.urlRequest, parameters: parameters)
 
         // Then
-        XCTAssertEqual(URLRequest.url?.query ?? "", "foo=1.1", "query is incorrect")
+        XCTAssertEqual(urlRequest.url?.query ?? "", "foo=1.1", "query is incorrect")
     }
 
     func testURLParameterEncodeStringKeyBoolValueParameter() {
@@ -122,10 +122,10 @@ class URLParameterEncodingTestCase: ParameterEncodingTestCase {
         let parameters = ["foo": true]
 
         // When
-        let (URLRequest, _) = encoding.encode(self.urlRequest, parameters: parameters)
+        let (urlRequest, _) = encoding.encode(self.urlRequest, parameters: parameters)
 
         // Then
-        XCTAssertEqual(URLRequest.url?.query ?? "", "foo=1", "query is incorrect")
+        XCTAssertEqual(urlRequest.url?.query ?? "", "foo=1", "query is incorrect")
     }
 
     func testURLParameterEncodeStringKeyArrayValueParameter() {
@@ -133,10 +133,10 @@ class URLParameterEncodingTestCase: ParameterEncodingTestCase {
         let parameters = ["foo": ["a", 1, true]]
 
         // When
-        let (URLRequest, _) = encoding.encode(self.urlRequest, parameters: parameters)
+        let (urlRequest, _) = encoding.encode(self.urlRequest, parameters: parameters)
 
         // Then
-        XCTAssertEqual(URLRequest.url?.query ?? "", "foo%5B%5D=a&foo%5B%5D=1&foo%5B%5D=1", "query is incorrect")
+        XCTAssertEqual(urlRequest.url?.query ?? "", "foo%5B%5D=a&foo%5B%5D=1&foo%5B%5D=1", "query is incorrect")
     }
 
     func testURLParameterEncodeStringKeyDictionaryValueParameter() {
@@ -144,10 +144,10 @@ class URLParameterEncodingTestCase: ParameterEncodingTestCase {
         let parameters = ["foo": ["bar": 1]]
 
         // When
-        let (URLRequest, _) = encoding.encode(self.urlRequest, parameters: parameters)
+        let (urlRequest, _) = encoding.encode(self.urlRequest, parameters: parameters)
 
         // Then
-        XCTAssertEqual(URLRequest.url?.query ?? "", "foo%5Bbar%5D=1", "query is incorrect")
+        XCTAssertEqual(urlRequest.url?.query ?? "", "foo%5Bbar%5D=1", "query is incorrect")
     }
 
     func testURLParameterEncodeStringKeyNestedDictionaryValueParameter() {
@@ -155,10 +155,10 @@ class URLParameterEncodingTestCase: ParameterEncodingTestCase {
         let parameters = ["foo": ["bar": ["baz": 1]]]
 
         // When
-        let (URLRequest, _) = encoding.encode(self.urlRequest, parameters: parameters)
+        let (urlRequest, _) = encoding.encode(self.urlRequest, parameters: parameters)
 
         // Then
-        XCTAssertEqual(URLRequest.url?.query ?? "", "foo%5Bbar%5D%5Bbaz%5D=1", "query is incorrect")
+        XCTAssertEqual(urlRequest.url?.query ?? "", "foo%5Bbar%5D%5Bbaz%5D=1", "query is incorrect")
     }
 
     func testURLParameterEncodeStringKeyNestedDictionaryArrayValueParameter() {
@@ -166,11 +166,11 @@ class URLParameterEncodingTestCase: ParameterEncodingTestCase {
         let parameters = ["foo": ["bar": ["baz": ["a", 1, true]]]]
 
         // When
-        let (URLRequest, _) = encoding.encode(self.urlRequest, parameters: parameters)
+        let (urlRequest, _) = encoding.encode(self.urlRequest, parameters: parameters)
 
         // Then
         let expectedQuery = "foo%5Bbar%5D%5Bbaz%5D%5B%5D=a&foo%5Bbar%5D%5Bbaz%5D%5B%5D=1&foo%5Bbar%5D%5Bbaz%5D%5B%5D=1"
-        XCTAssertEqual(URLRequest.url?.query ?? "", expectedQuery, "query is incorrect")
+        XCTAssertEqual(urlRequest.url?.query ?? "", expectedQuery, "query is incorrect")
     }
 
     // MARK: Tests - All Reserved / Unreserved / Illegal Characters According to RFC 3986
@@ -182,11 +182,11 @@ class URLParameterEncodingTestCase: ParameterEncodingTestCase {
         let parameters = ["reserved": "\(generalDelimiters)\(subDelimiters)"]
 
         // When
-        let (URLRequest, _) = encoding.encode(self.urlRequest, parameters: parameters)
+        let (urlRequest, _) = encoding.encode(self.urlRequest, parameters: parameters)
 
         // Then
         let expectedQuery = "reserved=%3A%23%5B%5D%40%21%24%26%27%28%29%2A%2B%2C%3B%3D"
-        XCTAssertEqual(URLRequest.url?.query ?? "", expectedQuery, "query is incorrect")
+        XCTAssertEqual(urlRequest.url?.query ?? "", expectedQuery, "query is incorrect")
     }
 
     func testThatReservedCharactersQuestionMarkAndForwardSlashAreNotPercentEscaped() {
@@ -194,10 +194,10 @@ class URLParameterEncodingTestCase: ParameterEncodingTestCase {
         let parameters = ["reserved": "?/"]
 
         // When
-        let (URLRequest, _) = encoding.encode(self.urlRequest, parameters: parameters)
+        let (urlRequest, _) = encoding.encode(self.urlRequest, parameters: parameters)
 
         // Then
-        XCTAssertEqual(URLRequest.url?.query ?? "", "reserved=?/", "query is incorrect")
+        XCTAssertEqual(urlRequest.url?.query ?? "", "reserved=?/", "query is incorrect")
     }
 
     func testThatUnreservedNumericCharactersAreNotPercentEscaped() {
@@ -205,10 +205,10 @@ class URLParameterEncodingTestCase: ParameterEncodingTestCase {
         let parameters = ["numbers": "0123456789"]
 
         // When
-        let (URLRequest, _) = encoding.encode(self.urlRequest, parameters: parameters)
+        let (urlRequest, _) = encoding.encode(self.urlRequest, parameters: parameters)
 
         // Then
-        XCTAssertEqual(URLRequest.url?.query ?? "", "numbers=0123456789", "query is incorrect")
+        XCTAssertEqual(urlRequest.url?.query ?? "", "numbers=0123456789", "query is incorrect")
     }
 
     func testThatUnreservedLowercaseCharactersAreNotPercentEscaped() {
@@ -216,10 +216,10 @@ class URLParameterEncodingTestCase: ParameterEncodingTestCase {
         let parameters = ["lowercase": "abcdefghijklmnopqrstuvwxyz"]
 
         // When
-        let (URLRequest, _) = encoding.encode(self.urlRequest, parameters: parameters)
+        let (urlRequest, _) = encoding.encode(self.urlRequest, parameters: parameters)
 
         // Then
-        XCTAssertEqual(URLRequest.url?.query ?? "", "lowercase=abcdefghijklmnopqrstuvwxyz", "query is incorrect")
+        XCTAssertEqual(urlRequest.url?.query ?? "", "lowercase=abcdefghijklmnopqrstuvwxyz", "query is incorrect")
     }
 
     func testThatUnreservedUppercaseCharactersAreNotPercentEscaped() {
@@ -227,10 +227,10 @@ class URLParameterEncodingTestCase: ParameterEncodingTestCase {
         let parameters = ["uppercase": "ABCDEFGHIJKLMNOPQRSTUVWXYZ"]
 
         // When
-        let (URLRequest, _) = encoding.encode(self.urlRequest, parameters: parameters)
+        let (urlRequest, _) = encoding.encode(self.urlRequest, parameters: parameters)
 
         // Then
-        XCTAssertEqual(URLRequest.url?.query ?? "", "uppercase=ABCDEFGHIJKLMNOPQRSTUVWXYZ", "query is incorrect")
+        XCTAssertEqual(urlRequest.url?.query ?? "", "uppercase=ABCDEFGHIJKLMNOPQRSTUVWXYZ", "query is incorrect")
     }
 
     func testThatIllegalASCIICharactersArePercentEscaped() {
@@ -238,11 +238,11 @@ class URLParameterEncodingTestCase: ParameterEncodingTestCase {
         let parameters = ["illegal": " \"#%<>[]\\^`{}|"]
 
         // When
-        let (URLRequest, _) = encoding.encode(self.urlRequest, parameters: parameters)
+        let (urlRequest, _) = encoding.encode(self.urlRequest, parameters: parameters)
 
         // Then
         let expectedQuery = "illegal=%20%22%23%25%3C%3E%5B%5D%5C%5E%60%7B%7D%7C"
-        XCTAssertEqual(URLRequest.url?.query ?? "", expectedQuery, "query is incorrect")
+        XCTAssertEqual(urlRequest.url?.query ?? "", expectedQuery, "query is incorrect")
     }
 
     // MARK: Tests - Special Character Queries
@@ -252,10 +252,10 @@ class URLParameterEncodingTestCase: ParameterEncodingTestCase {
         let parameters = ["foo&bar": "baz&qux", "foobar": "bazqux"]
 
         // When
-        let (URLRequest, _) = encoding.encode(self.urlRequest, parameters: parameters)
+        let (urlRequest, _) = encoding.encode(self.urlRequest, parameters: parameters)
 
         // Then
-        XCTAssertEqual(URLRequest.url?.query ?? "", "foo%26bar=baz%26qux&foobar=bazqux", "query is incorrect")
+        XCTAssertEqual(urlRequest.url?.query ?? "", "foo%26bar=baz%26qux&foobar=bazqux", "query is incorrect")
     }
 
     func testURLParameterEncodeStringWithQuestionMarkKeyStringWithQuestionMarkValueParameter() {
@@ -263,10 +263,10 @@ class URLParameterEncodingTestCase: ParameterEncodingTestCase {
         let parameters = ["?foo?": "?bar?"]
 
         // When
-        let (URLRequest, _) = encoding.encode(self.urlRequest, parameters: parameters)
+        let (urlRequest, _) = encoding.encode(self.urlRequest, parameters: parameters)
 
         // Then
-        XCTAssertEqual(URLRequest.url?.query ?? "", "?foo?=?bar?", "query is incorrect")
+        XCTAssertEqual(urlRequest.url?.query ?? "", "?foo?=?bar?", "query is incorrect")
     }
 
     func testURLParameterEncodeStringWithSlashKeyStringWithQuestionMarkValueParameter() {
@@ -274,10 +274,10 @@ class URLParameterEncodingTestCase: ParameterEncodingTestCase {
         let parameters = ["foo": "/bar/baz/qux"]
 
         // When
-        let (URLRequest, _) = encoding.encode(self.urlRequest, parameters: parameters)
+        let (urlRequest, _) = encoding.encode(self.urlRequest, parameters: parameters)
 
         // Then
-        XCTAssertEqual(URLRequest.url?.query ?? "", "foo=/bar/baz/qux", "query is incorrect")
+        XCTAssertEqual(urlRequest.url?.query ?? "", "foo=/bar/baz/qux", "query is incorrect")
     }
 
     func testURLParameterEncodeStringWithSpaceKeyStringWithSpaceValueParameter() {
@@ -285,10 +285,10 @@ class URLParameterEncodingTestCase: ParameterEncodingTestCase {
         let parameters = [" foo ": " bar "]
 
         // When
-        let (URLRequest, _) = encoding.encode(self.urlRequest, parameters: parameters)
+        let (urlRequest, _) = encoding.encode(self.urlRequest, parameters: parameters)
 
         // Then
-        XCTAssertEqual(URLRequest.url?.query ?? "", "%20foo%20=%20bar%20", "query is incorrect")
+        XCTAssertEqual(urlRequest.url?.query ?? "", "%20foo%20=%20bar%20", "query is incorrect")
     }
 
     func testURLParameterEncodeStringWithPlusKeyStringWithPlusValueParameter() {
@@ -296,10 +296,10 @@ class URLParameterEncodingTestCase: ParameterEncodingTestCase {
         let parameters = ["+foo+": "+bar+"]
 
         // When
-        let (URLRequest, _) = encoding.encode(self.urlRequest, parameters: parameters)
+        let (urlRequest, _) = encoding.encode(self.urlRequest, parameters: parameters)
 
         // Then
-        XCTAssertEqual(URLRequest.url?.query ?? "", "%2Bfoo%2B=%2Bbar%2B", "query is incorrect")
+        XCTAssertEqual(urlRequest.url?.query ?? "", "%2Bfoo%2B=%2Bbar%2B", "query is incorrect")
     }
 
     func testURLParameterEncodeStringKeyPercentEncodedStringValueParameter() {
@@ -307,10 +307,10 @@ class URLParameterEncodingTestCase: ParameterEncodingTestCase {
         let parameters = ["percent": "%25"]
 
         // When
-        let (URLRequest, _) = encoding.encode(self.urlRequest, parameters: parameters)
+        let (urlRequest, _) = encoding.encode(self.urlRequest, parameters: parameters)
 
         // Then
-        XCTAssertEqual(URLRequest.url?.query ?? "", "percent=%2525", "query is incorrect")
+        XCTAssertEqual(urlRequest.url?.query ?? "", "percent=%2525", "query is incorrect")
     }
 
     func testURLParameterEncodeStringKeyNonLatinStringValueParameter() {
@@ -323,7 +323,7 @@ class URLParameterEncodingTestCase: ParameterEncodingTestCase {
         ]
 
         // When
-        let (URLRequest, _) = encoding.encode(self.urlRequest, parameters: parameters)
+        let (urlRequest, _) = encoding.encode(self.urlRequest, parameters: parameters)
 
         // Then
         let expectedParameterValues = [
@@ -334,48 +334,48 @@ class URLParameterEncodingTestCase: ParameterEncodingTestCase {
         ]
 
         let expectedQuery = expectedParameterValues.joined(separator: "&")
-        XCTAssertEqual(URLRequest.url?.query ?? "", expectedQuery, "query is incorrect")
+        XCTAssertEqual(urlRequest.url?.query ?? "", expectedQuery, "query is incorrect")
     }
 
     func testURLParameterEncodeStringForRequestWithPrecomposedQuery() {
         // Given
-        let URL = Foundation.URL(string: "https://example.com/movies?hd=[1]")!
+        let url = URL(string: "https://example.com/movies?hd=[1]")!
         let parameters = ["page": "0"]
 
         // When
-        let (URLRequest, _) = encoding.encode(Foundation.URLRequest(url: URL), parameters: parameters)
+        let (urlRequest, _) = encoding.encode(URLRequest(url: url), parameters: parameters)
 
         // Then
-        XCTAssertEqual(URLRequest.url?.query ?? "", "hd=%5B1%5D&page=0", "query is incorrect")
+        XCTAssertEqual(urlRequest.url?.query ?? "", "hd=%5B1%5D&page=0", "query is incorrect")
     }
 
     func testURLParameterEncodeStringWithPlusKeyStringWithPlusValueParameterForRequestWithPrecomposedQuery() {
         // Given
-        let URL = Foundation.URL(string: "https://example.com/movie?hd=[1]")!
+        let url = URL(string: "https://example.com/movie?hd=[1]")!
         let parameters = ["+foo+": "+bar+"]
 
         // When
-        let (URLRequest, _) = encoding.encode(Foundation.URLRequest(url: URL), parameters: parameters)
+        let (urlRequest, _) = encoding.encode(URLRequest(url: url), parameters: parameters)
 
         // Then
-        XCTAssertEqual(URLRequest.url?.query ?? "", "hd=%5B1%5D&%2Bfoo%2B=%2Bbar%2B", "query is incorrect")
+        XCTAssertEqual(urlRequest.url?.query ?? "", "hd=%5B1%5D&%2Bfoo%2B=%2Bbar%2B", "query is incorrect")
     }
 
     func testURLParameterEncodeStringWithThousandsOfChineseCharacters() {
         // Given
         let repeatedCount = 2_000
-        let URL = Foundation.URL(string: "https://example.com/movies")!
+        let url = URL(string: "https://example.com/movies")!
         let parameters = ["chinese": String(count: repeatedCount, repeatedString: "一二三四五六七八九十")]
 
         // When
-        let (URLRequest, _) = encoding.encode(Foundation.URLRequest(url: URL), parameters: parameters)
+        let (urlRequest, _) = encoding.encode(URLRequest(url: url), parameters: parameters)
 
         // Then
         var expected = "chinese="
         for _ in 0..<repeatedCount {
             expected += "%E4%B8%80%E4%BA%8C%E4%B8%89%E5%9B%9B%E4%BA%94%E5%85%AD%E4%B8%83%E5%85%AB%E4%B9%9D%E5%8D%81"
         }
-        XCTAssertEqual(URLRequest.url?.query ?? "", expected, "query is incorrect")
+        XCTAssertEqual(urlRequest.url?.query ?? "", expected, "query is incorrect")
     }
 
     // MARK: Tests - Varying HTTP Methods
@@ -402,18 +402,19 @@ class URLParameterEncodingTestCase: ParameterEncodingTestCase {
         let parameters = ["foo": 1, "bar": 2]
 
         // When
-        let (URLRequest, _) = encoding.encode(mutableURLRequest, parameters: parameters)
+        let (urlRequest, _) = encoding.encode(mutableURLRequest, parameters: parameters)
 
         // Then
         XCTAssertEqual(
-            URLRequest.value(forHTTPHeaderField: "Content-Type") ?? "",
+            urlRequest.value(forHTTPHeaderField: "Content-Type") ?? "",
             "application/x-www-form-urlencoded; charset=utf-8",
             "Content-Type should be application/x-www-form-urlencoded"
         )
-        XCTAssertNotNil(URLRequest.httpBody, "HTTPBody should not be nil")
+        XCTAssertNotNil(urlRequest.httpBody, "HTTPBody should not be nil")
 
-        if let httpBody = URLRequest.httpBody,
-           let decodedHTTPBody = String(data: httpBody, encoding: String.Encoding.utf8)
+        if
+            let httpBody = urlRequest.httpBody,
+            let decodedHTTPBody = String(data: httpBody, encoding: String.Encoding.utf8)
         {
             XCTAssertEqual(decodedHTTPBody, "bar=2&foo=1", "HTTPBody is incorrect")
         } else {
@@ -428,12 +429,12 @@ class URLParameterEncodingTestCase: ParameterEncodingTestCase {
         let parameters = ["foo": 1, "bar": 2]
 
         // When
-        let (URLRequest, _) = ParameterEncoding.urlEncodedInURL.encode(mutableURLRequest, parameters: parameters)
+        let (urlRequest, _) = ParameterEncoding.urlEncodedInURL.encode(mutableURLRequest, parameters: parameters)
 
         // Then
-        XCTAssertEqual(URLRequest.url?.query ?? "", "bar=2&foo=1", "query is incorrect")
-        XCTAssertNil(URLRequest.value(forHTTPHeaderField: "Content-Type"), "Content-Type should be nil")
-        XCTAssertNil(URLRequest.httpBody, "HTTPBody should be nil")
+        XCTAssertEqual(urlRequest.url?.query ?? "", "bar=2&foo=1", "query is incorrect")
+        XCTAssertNil(urlRequest.value(forHTTPHeaderField: "Content-Type"), "Content-Type should be nil")
+        XCTAssertNil(urlRequest.httpBody, "HTTPBody should be nil")
     }
 }
 
@@ -503,18 +504,18 @@ class JSONParameterEncodingTestCase: ParameterEncodingTestCase {
 
     func testJSONParameterEncodeParametersRetainsCustomContentType() {
         // Given
-        var mutableURLRequest = Foundation.URLRequest(url: URL(string: "https://example.com/")!)
+        var mutableURLRequest = URLRequest(url: URL(string: "https://example.com/")!)
         mutableURLRequest.setValue("application/custom-json-type+json", forHTTPHeaderField: "Content-Type")
 
         let parameters = ["foo": "bar"]
 
         // When
-        let (URLRequest, error) = encoding.encode(mutableURLRequest, parameters: parameters)
+        let (urlRequest, error) = encoding.encode(mutableURLRequest, parameters: parameters)
 
         // Then
         XCTAssertNil(error)
-        XCTAssertNil(URLRequest.url?.query)
-        XCTAssertEqual(URLRequest.value(forHTTPHeaderField: "Content-Type"), "application/custom-json-type+json")
+        XCTAssertNil(urlRequest.url?.query)
+        XCTAssertEqual(urlRequest.value(forHTTPHeaderField: "Content-Type"), "application/custom-json-type+json")
     }
 }
 
@@ -626,7 +627,7 @@ class PropertyListParameterEncodingTestCase: ParameterEncodingTestCase {
 
     func testPropertyListParameterEncodeParametersRetainsCustomContentType() {
         // Given
-        var mutableURLRequest = Foundation.URLRequest(url: URL(string: "https://example.com/")!)
+        var mutableURLRequest = URLRequest(url: URL(string: "https://example.com/")!)
         mutableURLRequest.setValue("application/custom-plist-type+plist", forHTTPHeaderField: "Content-Type")
 
         let parameters = ["foo": "bar"]
@@ -648,7 +649,7 @@ class CustomParameterEncodingTestCase: ParameterEncodingTestCase {
 
     func testCustomParameterEncode() {
         // Given
-        let encodingClosure: (URLRequestConvertible, [String: AnyObject]?) -> (Foundation.URLRequest, NSError?) = { urlRequest, parameters in
+        let encodingClosure: (URLRequestConvertible, [String: AnyObject]?) -> (URLRequest, NSError?) = { urlRequest, parameters in
             guard let parameters = parameters else { return (urlRequest.urlRequest, nil) }
 
             var urlString = urlRequest.urlRequest.urlString + "?"
@@ -656,7 +657,7 @@ class CustomParameterEncodingTestCase: ParameterEncodingTestCase {
             parameters.forEach { urlString += "\($0)=\($1)" }
 
             var mutableURLRequest = urlRequest.urlRequest
-            mutableURLRequest.url = Foundation.URL(string: urlString)!
+            mutableURLRequest.url = URL(string: urlString)!
 
             return (mutableURLRequest, nil)
         }
@@ -665,8 +666,8 @@ class CustomParameterEncodingTestCase: ParameterEncodingTestCase {
         let encoding: ParameterEncoding = .custom(encodingClosure)
 
         // Then
-        let url = Foundation.URL(string: "https://example.com")!
-        let urlRequest = Foundation.URLRequest(url: url)
+        let url = URL(string: "https://example.com")!
+        let urlRequest = URLRequest(url: url)
         let parameters: [String: AnyObject] = ["foo": "bar"]
 
         let result = encoding.encode(urlRequest, parameters: parameters)
