+++ Tests/SessionManagerTests.swift
@@ -43,7 +43,7 @@ class SessionManagerTestCase: BaseTestCase {
     func testInitializerWithSpecifiedArguments() {
         // Given
         let configuration = URLSessionConfiguration.default
-        let delegate = SessionManager.SessionDelegate()
+        let delegate = SessionDelegate()
         let serverTrustPolicyManager = ServerTrustPolicyManager(policies: [:])
 
         // When
@@ -61,7 +61,7 @@ class SessionManagerTestCase: BaseTestCase {
 
     func testThatFailableInitializerSucceedsWithDefaultArguments() {
         // Given
-        let delegate = SessionManager.SessionDelegate()
+        let delegate = SessionDelegate()
         let session: URLSession = {
             let configuration = URLSessionConfiguration.default
             return URLSession(configuration: configuration, delegate: delegate, delegateQueue: nil)
@@ -81,7 +81,7 @@ class SessionManagerTestCase: BaseTestCase {
 
     func testThatFailableInitializerSucceedsWithSpecifiedArguments() {
         // Given
-        let delegate = SessionManager.SessionDelegate()
+        let delegate = SessionDelegate()
         let session: URLSession = {
             let configuration = URLSessionConfiguration.default
             return URLSession(configuration: configuration, delegate: delegate, delegateQueue: nil)
@@ -103,10 +103,10 @@ class SessionManagerTestCase: BaseTestCase {
 
     func testThatFailableInitializerFailsWithWhenDelegateDoesNotEqualSessionDelegate() {
         // Given
-        let delegate = SessionManager.SessionDelegate()
+        let delegate = SessionDelegate()
         let session: URLSession = {
             let configuration = URLSessionConfiguration.default
-            return URLSession(configuration: configuration, delegate: SessionManager.SessionDelegate(), delegateQueue: nil)
+            return URLSession(configuration: configuration, delegate: SessionDelegate(), delegateQueue: nil)
         }()
 
         // When
@@ -118,7 +118,7 @@ class SessionManagerTestCase: BaseTestCase {
 
     func testThatFailableInitializerFailsWhenSessionDelegateIsNil() {
         // Given
-        let delegate = SessionManager.SessionDelegate()
+        let delegate = SessionDelegate()
         let session: URLSession = {
             let configuration = URLSessionConfiguration.default
             return URLSession(configuration: configuration, delegate: nil, delegateQueue: nil)
@@ -138,15 +138,15 @@ class SessionManagerTestCase: BaseTestCase {
         let manager = SessionManager()
         manager.startRequestsImmediately = false
 
-        let URL = Foundation.URL(string: "https://httpbin.org/get")!
-        let URLRequest = Foundation.URLRequest(url: URL)
+        let url = URL(string: "https://httpbin.org/get")!
+        let urlRequest = URLRequest(url: url)
 
-        let expectation = self.expectation(description: "\(URL)")
+        let expectation = self.expectation(description: "\(url)")
 
         var response: HTTPURLResponse?
 
         // When
-        manager.request(URLRequest)
+        manager.dataRequest(urlRequest: urlRequest)
             .response { _, responseResponse, _, _ in
                 response = responseResponse
                 expectation.fulfill()
@@ -167,11 +167,11 @@ class SessionManagerTestCase: BaseTestCase {
         var manager: SessionManager? = SessionManager()
         manager?.startRequestsImmediately = false
 
-        let URL = Foundation.URL(string: "https://httpbin.org/get")!
-        let URLRequest = Foundation.URLRequest(url: URL)
+        let url = URL(string: "https://httpbin.org/get")!
+        let urlRequest = URLRequest(url: url)
 
         // When
-        let request = manager?.request(URLRequest)
+        let request = manager?.dataRequest(urlRequest: urlRequest)
         manager = nil
 
         // Then
@@ -184,11 +184,11 @@ class SessionManagerTestCase: BaseTestCase {
         var manager: SessionManager? = SessionManager()
         manager!.startRequestsImmediately = false
 
-        let URL = Foundation.URL(string: "https://httpbin.org/get")!
-        let URLRequest = Foundation.URLRequest(url: URL)
+        let url = URL(string: "https://httpbin.org/get")!
+        let urlRequest = URLRequest(url: url)
 
         // When
-        let request = manager!.request(URLRequest)
+        let request = manager!.dataRequest(urlRequest: urlRequest)
         request.cancel()
         manager = nil
 
@@ -253,7 +253,7 @@ class ManagerConfigurationHeadersTestCase: BaseTestCase {
         var response: Response<AnyObject, NSError>?
 
         // When
-        manager.request(.GET, "https://httpbin.org/headers")
+        manager.dataRequest(method: .GET, urlString: "https://httpbin.org/headers")
             .responseJSON { closureResponse in
                 response = closureResponse
                 expectation.fulfill()
