+++ Tests/TLSEvaluationTests.swift
@@ -33,8 +33,8 @@ private struct TestCertificates {
     static let Leaf = TestCertificates.certificateWithFileName("expired.badssl.com-leaf")
 
     static func certificateWithFileName(_ fileName: String) -> SecCertificate {
-        class Bundle {}
-        let filePath = Foundation.Bundle(for: Bundle.self).path(forResource: fileName, ofType: "cer")!
+        class Locater {}
+        let filePath = Bundle(for: Locater.self).path(forResource: fileName, ofType: "cer")!
         let data = try! Data(contentsOf: URL(fileURLWithPath: filePath))
         let certificate = SecCertificateCreateWithData(nil, data)!
 
@@ -64,7 +64,7 @@ private struct TestPublicKeys {
 // MARK: -
 
 class TLSEvaluationExpiredLeafCertificateTestCase: BaseTestCase {
-    let URL = "https://expired.badssl.com/"
+    let urlString = "https://expired.badssl.com/"
     let host = "expired.badssl.com"
     var configuration: URLSessionConfiguration!
 
@@ -79,12 +79,12 @@ class TLSEvaluationExpiredLeafCertificateTestCase: BaseTestCase {
 
     func testThatExpiredCertificateRequestFailsWithNoServerTrustPolicy() {
         // Given
-        weak var expectation = self.expectation(description: "\(URL)")
+        weak var expectation = self.expectation(description: "\(urlString)")
         let manager = SessionManager(configuration: configuration)
         var error: NSError?
 
         // When
-        manager.request(.GET, URL)
+        manager.dataRequest(method: .GET, urlString: urlString)
             .response { _, _, _, responseError in
                 error = responseError
                 expectation?.fulfill()
@@ -112,11 +112,11 @@ class TLSEvaluationExpiredLeafCertificateTestCase: BaseTestCase {
             serverTrustPolicyManager: ServerTrustPolicyManager(policies: policies)
         )
 
-        weak var expectation = self.expectation(description: "\(URL)")
+        weak var expectation = self.expectation(description: "\(urlString)")
         var error: NSError?
 
         // When
-        manager.request(.GET, URL)
+        manager.dataRequest(method: .GET, urlString: urlString)
             .response { _, _, _, responseError in
                 error = responseError
                 expectation?.fulfill()
@@ -148,11 +148,11 @@ class TLSEvaluationExpiredLeafCertificateTestCase: BaseTestCase {
             serverTrustPolicyManager: ServerTrustPolicyManager(policies: policies)
         )
 
-        weak var expectation = self.expectation(description: "\(URL)")
+        weak var expectation = self.expectation(description: "\(urlString)")
         var error: NSError?
 
         // When
-        manager.request(.GET, URL)
+        manager.dataRequest(method: .GET, urlString: urlString)
             .response { _, _, _, responseError in
                 error = responseError
                 expectation?.fulfill()
@@ -188,11 +188,11 @@ class TLSEvaluationExpiredLeafCertificateTestCase: BaseTestCase {
             serverTrustPolicyManager: ServerTrustPolicyManager(policies: policies)
         )
 
-        weak var expectation = self.expectation(description: "\(URL)")
+        weak var expectation = self.expectation(description: "\(urlString)")
         var error: NSError?
 
         // When
-        manager.request(.GET, URL)
+        manager.dataRequest(method: .GET, urlString: urlString)
             .response { _, _, _, responseError in
                 error = responseError
                 expectation?.fulfill()
@@ -222,11 +222,11 @@ class TLSEvaluationExpiredLeafCertificateTestCase: BaseTestCase {
             serverTrustPolicyManager: ServerTrustPolicyManager(policies: policies)
         )
 
-        weak var expectation = self.expectation(description: "\(URL)")
+        weak var expectation = self.expectation(description: "\(urlString)")
         var error: NSError?
 
         // When
-        manager.request(.GET, URL)
+        manager.dataRequest(method: .GET, urlString: urlString)
             .response { _, _, _, responseError in
                 error = responseError
                 expectation?.fulfill()
@@ -250,11 +250,11 @@ class TLSEvaluationExpiredLeafCertificateTestCase: BaseTestCase {
             serverTrustPolicyManager: ServerTrustPolicyManager(policies: policies)
         )
 
-        weak var expectation = self.expectation(description: "\(URL)")
+        weak var expectation = self.expectation(description: "\(urlString)")
         var error: NSError?
 
         // When
-        manager.request(.GET, URL)
+        manager.dataRequest(method: .GET, urlString: urlString)
             .response { _, _, _, responseError in
                 error = responseError
                 expectation?.fulfill()
@@ -278,11 +278,11 @@ class TLSEvaluationExpiredLeafCertificateTestCase: BaseTestCase {
             serverTrustPolicyManager: ServerTrustPolicyManager(policies: policies)
         )
 
-        weak var expectation = self.expectation(description: "\(URL)")
+        weak var expectation = self.expectation(description: "\(urlString)")
         var error: NSError?
 
         // When
-        manager.request(.GET, URL)
+        manager.dataRequest(method: .GET, urlString: urlString)
             .response { _, _, _, responseError in
                 error = responseError
                 expectation?.fulfill()
@@ -308,11 +308,11 @@ class TLSEvaluationExpiredLeafCertificateTestCase: BaseTestCase {
             serverTrustPolicyManager: ServerTrustPolicyManager(policies: policies)
         )
 
-        weak var expectation = self.expectation(description: "\(URL)")
+        weak var expectation = self.expectation(description: "\(urlString)")
         var error: NSError?
 
         // When
-        manager.request(.GET, URL)
+        manager.dataRequest(method: .GET, urlString: urlString)
             .response { _, _, _, responseError in
                 error = responseError
                 expectation?.fulfill()
@@ -342,11 +342,11 @@ class TLSEvaluationExpiredLeafCertificateTestCase: BaseTestCase {
             serverTrustPolicyManager: ServerTrustPolicyManager(policies: policies)
         )
 
-        weak var expectation = self.expectation(description: "\(URL)")
+        weak var expectation = self.expectation(description: "\(urlString)")
         var error: NSError?
 
         // When
-        manager.request(.GET, URL)
+        manager.dataRequest(method: .GET, urlString: urlString)
             .response { _, _, _, responseError in
                 error = responseError
                 expectation?.fulfill()
@@ -370,11 +370,11 @@ class TLSEvaluationExpiredLeafCertificateTestCase: BaseTestCase {
             serverTrustPolicyManager: ServerTrustPolicyManager(policies: policies)
         )
 
-        weak var expectation = self.expectation(description: "\(URL)")
+        weak var expectation = self.expectation(description: "\(urlString)")
         var error: NSError?
 
         // When
-        manager.request(.GET, URL)
+        manager.dataRequest(method: .GET, urlString: urlString)
             .response { _, _, _, responseError in
                 error = responseError
                 expectation?.fulfill()
@@ -398,11 +398,11 @@ class TLSEvaluationExpiredLeafCertificateTestCase: BaseTestCase {
             serverTrustPolicyManager: ServerTrustPolicyManager(policies: policies)
         )
 
-        weak var expectation = self.expectation(description: "\(URL)")
+        weak var expectation = self.expectation(description: "\(urlString)")
         var error: NSError?
 
         // When
-        manager.request(.GET, URL)
+        manager.dataRequest(method: .GET, urlString: urlString)
             .response { _, _, _, responseError in
                 error = responseError
                 expectation?.fulfill()
@@ -424,11 +424,11 @@ class TLSEvaluationExpiredLeafCertificateTestCase: BaseTestCase {
             serverTrustPolicyManager: ServerTrustPolicyManager(policies: policies)
         )
 
-        weak var expectation = self.expectation(description: "\(URL)")
+        weak var expectation = self.expectation(description: "\(urlString)")
         var error: NSError?
 
         // When
-        manager.request(.GET, URL)
+        manager.dataRequest(method: .GET, urlString: urlString)
             .response { _, _, _, responseError in
                 error = responseError
                 expectation?.fulfill()
@@ -456,11 +456,11 @@ class TLSEvaluationExpiredLeafCertificateTestCase: BaseTestCase {
             serverTrustPolicyManager: ServerTrustPolicyManager(policies: policies)
         )
 
-        weak var expectation = self.expectation(description: "\(URL)")
+        weak var expectation = self.expectation(description: "\(urlString)")
         var error: NSError?
 
         // When
-        manager.request(.GET, URL)
+        manager.dataRequest(method: .GET, urlString: urlString)
             .response { _, _, _, responseError in
                 error = responseError
                 expectation?.fulfill()
@@ -486,11 +486,11 @@ class TLSEvaluationExpiredLeafCertificateTestCase: BaseTestCase {
             serverTrustPolicyManager: ServerTrustPolicyManager(policies: policies)
         )
 
-        weak var expectation = self.expectation(description: "\(URL)")
+        weak var expectation = self.expectation(description: "\(urlString)")
         var error: NSError?
 
         // When
-        manager.request(.GET, URL)
+        manager.dataRequest(method: .GET, urlString: urlString)
             .response { _, _, _, responseError in
                 error = responseError
                 expectation?.fulfill()
